backend/requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and security
bcrypt==4.1.2
pyjwt==2.8.0
email-validator==2.1.0

# Validation and serialization
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
factory-boy==3.3.0